dashboard/frontend/requirements.txt
streamlit
streamlit_autorefresh
websocket-client
plotly
numpy
pandas
streamlit-plotly-events
